requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
exceptiongroup==1.3.1
greenlet==3.2.4
idna==3.11
iniconfig==2.1.0
packaging==26.0
playwright==1.58.0
pluggy==1.6.0
pyee==13.0.1
Pygments==2.19.2
pytest==8.4.2
pytest-base-url==2.1.0
pytest-playwright==0.7.1
python-slugify==8.0.4
requests==2.32.5
text-unidecode==1.3
tomli==2.4.0
typing_extensions==4.15.0
urllib3==2.6.3
allure-pytest

# git clone repo
# python3 -m venv venv
# source venv/bin/activate
# pip install -r requirements.txt
# python -m playwright install
# pytest

